x86/sev: Remap the SVSM CAA area using SVSM_CORE_REMAP_CAA call
Since the firmware supplied SVSM CAA for the BSP is likely to be in
reserved memory, switch off it to a kernel SVSM CAA so that access and
use of the CAA is available during the boot.

Once it is safe to allocate/reserve memory, allocate an SVSM calling area
for each vCPU. After allocating the calling area, the BSP will switch to
the new area immediately. The calling area for an AP is identified to the
SVSM when creating the VMSA in preparation for booting the AP.

Merged with host kernel changes: rename SVM_VMGEXIT_SNP_RUN_VMPL
to SVM_VMGEXIT_RUN_VMPL: Roy Hopkins <rhopkins@suse.de>

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
Signed-off-by: Roy Hopkins <rhopkins@suse.de>